Add Linux support for CLI screenshot renderer
- Add fontconfig-based font manager for Linux in IGraphicsSkia
- Add HeadlessFont class for font loading on non-Apple platforms
- Add IGRAPHICS_HEADLESS platform check in include header
- Add FONT_DESCRIPTOR_TYPE definition for Linux
- Add SetResourceBasePath/GetResourceBasePath for CLI resource paths
- Enhance Linux LocateResource to search fonts/ and img/ subdirectories
- Fix static library link order in CLI-Graphics.cmake (libskia.a last)
- Fix IGraphicsLinux.h syntax errors (destructor, class semicolon)
- Fix std::fmodf/std::expf to std::fmod/std::exp for GCC compatibility

+++ IGraphics/IGraphics.cpp
@@ -2171,7 +2171,7 @@ void IGraphics::ApplyLayerDropShadow(ILayerPtr& layer, const IShadow& shadow)
   kernel.Resize(iSize);
         
   for (int i = 0; i < iSize; i++)
-    kernel.Get()[i] = static_cast<uint8_t>(std::round(255.f * std::expf(-(i * i) * blurConst)));
+    kernel.Get()[i] = static_cast<uint8_t>(std::round(255.f * std::exp(-(i * i) * blurConst)));
   
   // Kernel normalisation
   int normFactor = kernel.Get()[0];
